fix: TEXT token position value
changes:
    - fix TEXT token position value
    - add BANG token ('!')
    - fix TEXT token adding

Previously all positions of TEXT tokens were incorrect, due to the fact that
all characters until a special symbol were read in and the position at
which the scanner stops is recorded in the token structure. The simple
fix is to subtract the lenght of the recorded text value of its
position.