refactor: reworked cli parsing
it is now easier to add a new cli flag. The new implementation automatically adds any flag of the cli.Flag structure which was added to the cli.OPTIONS array to the flag parsing logic. It also supports displaying all added options with their name and description to the help page which can be viewed by running fleck without any arguments.