fix(parser): code() dash parsing bug
due to the architecture of the inline code element parsing the code() function failed to parse inline code elements containing any special markdown character at the start of the element. This issue occured due to the fact, that i assumed a BACKTICK (`) always has the TEXT as a follow element which was in fact not correct and in retroperspective seems dumb. Whatever, its fixed now!