feat: watch mode '--watch'
if fleck is invoked with the '--watch' option, it will compile the file
and afterwards waits until the file size or the file last modification
time differs from the last check. If the before is true, fleck will
recompile the source file and wait for a new change. fleck will perform this check every
second.

changes:
- added '--watch' command
- split main.main into main.main and main.run
- add new main.watchForChanges function
- use the os.Stat function to watch for file changes
- changed log wording in the main.run function