feat(core): open tab in default browser in live-mode
starting the livemode using 'fleck --live-preview <file.md>' now issues
a new command via the 'os/exec.Command' function which spawns / opens a
new tab in the browser set as default. The url of the tab is pointed to
the location of the served html file. By default:
'http://localhost:12345/<file.html>'