fix: index error if source file is a directory
This fixes a crash caused by invoking fleck with a directory:

    before:
        $ fleck .
        panic: runtime error: index out of range [0] with length 0

        goroutine 6 [running]:
        github.com/xnacly/fleck/scanner.New({0x7ffd9b38733c?, 0x0?})
                /home/teo/programming/fleck/scanner/scanner.go:37 +0x1bb
        github.com/xnacly/fleck/core.Run({0x7ffd9b38733c, 0x1})
                /home/teo/programming/fleck/core/core.go:162 +0x387
        main.main.func1({0x7ffd9b38733c, 0x1})
                /home/teo/programming/fleck/fleck.go:82 +0x3e5
        created by main.main
                /home/teo/programming/fleck/fleck.go:56 +0x458
    now:
        $ fleck .
        2023/05/19 07:29:12 error: '.' is a directory
        exit status 1

This change also changes behaviour for empty files. When encountering
them, fleck now simply ignores them instead of throwing an error:

    before:
        $ fleck tester.md
        2023/05/19 07:29:55 warn: file is empty, exiting.

    now:
        $ fleck tester.md
        2023/05/19 07:31:18 info: detected empty source file (tester.md), skipping

Previously fleck only prints the following message when encountering a
file which can't be resolved, this error message was replaced with a
more in depth information on why the file couldn't be resolved:

    before:
        $ fleck test
        2023/05/19 07:32:45 error: failed to stat the file

    now:
        $ fleck test
        2023/05/19 07:33:34 error: stat test: no such file or directory

BREAKING CHANGE: this commit changes behaviour for passing empty source
files to fleck, as well as fixes a critical out of bounds error when
calling fleck with a directory.